Add components which receive education entry info
Adds two components to take in user input for education entries. One
component will act as a container for every entry form and will also
allow the creation of new entries. The other component is the form for
each individual entry and as updates entry data as well as allows the
deletion of the entire entry.