Add component for Practical Experience section
Adds component which will display the Practical Experience section of
the cv application. It holds the 3 state variables: the display boolean,
the list of practical experience entries, and an object which holds all
the existing responsibilities that are created for every entry in the
previous state variable. These state variables will be used in its two
child components.